Add English translations to kitsas_en.ts
Small commit to add English translations to kitsas_en.ts.
I'm not an accountant, and didn't check correctness of all the translations.
Also fix a typo in uusitilikartta.ui and add kitsas_en.ts to tr.qrc.
Partially fixes #1285.
More work is needed for full English translation of the app.